requirements.txt
BeautifulSoup
Django<1.7
Markdown
MySQL-python
South
Werkzeug
argparse
django-debug-toolbar
django-extensions
django-crispy-forms
django-picklefield
django-tracking2
django-tastypie
django-bootstrap3
django-mobile
gunicorn
slumber
ipython
jsonfield
requests
simplejson
six
sqlparse
uWSGI
wsgiref
celery[redis]
#mailsnake
boto
django-daterange-filter
django-dashing
django-social-auth
requests-toolbelt
pandas
django-js-reverse
pyExcelerator
requests-futures
